Output sample logs from fitting and GSASII tab
Move output_sample_logs and workspace_record into common class files
