+++ qt/python/mantidqtinterfaces/mantidqtinterfaces/Engineering/gui/engineering_diffraction/tabs/fitting/data_handling/data_model.py
@@ -4,146 +4,31 @@
 #   NScD Oak Ridge National Laboratory, European Spallation Source,
 #   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 # SPDX - License - Identifier: GPL - 3.0 +
-from os import path
 
-from mantid.simpleapi import Load, logger, EnggEstimateFocussedBackground, Minus, AverageLogData, SetUncertainties, \
-    CreateEmptyTableWorkspace, GroupWorkspaces, DeleteWorkspace, DeleteTableRows, RenameWorkspace, CreateWorkspace
+from mantid.simpleapi import Load, logger, CreateEmptyTableWorkspace, GroupWorkspaces, DeleteWorkspace
 from mantidqtinterfaces.Engineering.gui.engineering_diffraction.settings.settings_helper import get_setting
 from mantidqtinterfaces.Engineering.gui.engineering_diffraction.tabs.common import output_settings
+from mantidqtinterfaces.Engineering.gui.engineering_diffraction.tabs.common.output_sample_logs import \
+    SampleLogsGroupWorkspace, write_table_row, _generate_workspace_name
+from mantidqtinterfaces.Engineering.gui.engineering_diffraction.tabs.common.workspace_record import \
+    FittingWorkspaceRecordContainer
 from mantid.api import AnalysisDataService as ADS
 from mantid.api import TextAxis
 from mantid.kernel import UnitConversion, DeltaEModeType, UnitParams
 from matplotlib.pyplot import subplots
 from numpy import full, nan, max, array, vstack, argsort
 from itertools import chain
-from collections import defaultdict, OrderedDict
+from collections import defaultdict
 from re import findall, sub
-
-
-class FittingWorkspaceRecord:
-    """ Record that maps the workspace the user has loaded to the derived background-subtracted workspace"""
-    def __init__(self, **kwargs):
-        self.loaded_ws = kwargs.get('loaded_ws', None)
-        self.bgsub_ws_name = kwargs.get('bgsub_ws_name', None)
-        self.bgsub_ws = kwargs.get('bgsub_ws', None)
-        self.bg_params = kwargs.get('bg_params', [])
-
-    def get_bg_active(self):
-        if self.bgsub_ws and self.bg_params[0]:
-            # first element is isSub checkbox
-            return True
-        else:
-            return False
-
-    def get_active_ws(self):
-        if self.get_bg_active():
-            return self.bgsub_ws
-        else:
-            return self.loaded_ws
-
-    def __setattr__(self, key, value):
-        if key == 'bg_params':
-            if not isinstance(value, list):
-                raise AttributeError("bg_params must be a list")
-
-        self.__dict__[key] = value  # initialize self.key
-
-
-class FittingWorkspaceRecordContainer:
-    def __init__(self):
-        self.dict = OrderedDict()
-
-    def __getitem__(self, key):
-        return self.dict[key]
-
-    def __len__(self):
-        return len(self.dict)
-
-    def __bool__(self):
-        return len(self.dict)>0
-
-    def get(self, key, default_value):
-        return self.dict.get(key, default_value)
-
-    def add(self, ws_name, **kwargs):
-        self.dict[ws_name] = FittingWorkspaceRecord(**kwargs)
-
-    def add_from_names_dict(self, names_dict):
-        for key, value in names_dict.items():
-            self.add(key, bgsub_ws_name=value[0], bg_params=value[1])
-
-    def get_loaded_workpace_names(self):
-        return list(self.dict.keys())
-
-    def get_bgsub_workpace_names(self):
-        return [w.bgsub_ws_name for w in self.dict.values()]
-
-    # Set of methods that each return a two column dictionary for the various fields in FittingWorkspaceRecord
-    def get_loaded_ws_dict(self):
-        return dict([(key, value.loaded_ws) for key, value in self.dict.items()])
-
-    def get_bgsub_ws_dict(self):
-        return dict([(key, value.bgsub_ws) for key, value in self.dict.items()])
-
-    def get_bgsub_ws_name_dict(self):
-        return dict([(key, value.bgsub_ws_name) for key, value in self.dict.items()])
-
-    def get_bg_params_dict(self):
-        return dict([(key, value.bg_params) for key, value in self.dict.items()])
-
-    def get_active_ws_name_list(self):
-        return [self.get_active_ws_name(key) for key, value in self.dict.items()]
-
-    def get_active_ws_dict(self):
-        return dict([(self.get_active_ws_name(key), value.get_active_ws()) for key, value in self.dict.items()])
-
-    def get_ws_names_dict(self):
-        return dict([(key, [value.bgsub_ws_name, value.bg_params]) for key, value in self.dict.items()])
-
-    def get_loaded_workspace_name_from_bgsub(self, bgsub_ws_name):
-        return next((key for key, val in self.dict.items() if val.bgsub_ws_name == bgsub_ws_name), None)
-
-    def get_active_ws_name(self, ws_name):
-        ws_rec = self.dict[ws_name]
-        if ws_rec.get_bg_active():
-            return ws_rec.bgsub_ws_name
-        else:
-            return ws_name
-
-    def rename(self, old_ws_name, new_ws_name):
-        ws_loaded = self.dict.get(old_ws_name, None)
-        if ws_loaded:
-            self.dict[new_ws_name] = self.pop(old_ws_name)
-        else:
-            ws_loaded_name = self.get_loaded_workspace_name_from_bgsub(old_ws_name)
-            if ws_loaded_name:
-                self.dict[ws_loaded_name].bgsub_ws_name = new_ws_name
-
-    def replace_workspace(self, name, workspace):
-        ws_loaded = self.dict.get(name, None)
-        if ws_loaded:
-            self.dict[name].loaded_ws = workspace
-        else:
-            ws_loaded_name = self.get_loaded_workspace_name_from_bgsub(name)
-            if ws_loaded_name:
-                self.dict[ws_loaded_name].bgsub_ws = workspace
-
-    def pop(self, ws_name):
-        return self.dict.pop(ws_name)
-
-    def clear(self):
-        self.dict.clear()
 
 
 class FittingDataModel(object):
     def __init__(self):
-        self._log_names = []
-        self._log_workspaces = None # GroupWorkspace
-        self._log_values = dict()  # {ws_name: {log_name: [avg, er]} }
         self._fit_results = {}  # {WorkspaceName: fit_result_dict}
         self._fit_workspaces = None
         self._last_added = []  # List of workspace names loaded in the last load action.
         self._data_workspaces = FittingWorkspaceRecordContainer()
+        self._sample_logs_workspace_group = SampleLogsGroupWorkspace()
 
     def restore_files(self, ws_names):
         self._data_workspaces.add_from_names_dict(ws_names)
@@ -163,13 +48,13 @@ class FittingDataModel(object):
             except RuntimeError as e:
                 logger.error(
                     f"Failed to restore workspace: {ws_name}. Error: {e}. \n Continuing loading of other files.")
-        self.update_log_workspace_group()
+        self._sample_logs_workspace_group.update_log_workspace_group(self._data_workspaces)
 
     def load_files(self, filenames_string):
         self._last_added = []
         filenames = [name.strip() for name in filenames_string.split(",")]
         for filename in filenames:
-            ws_name = self._generate_workspace_name(filename)
+            ws_name = _generate_workspace_name(filename)
             if ws_name not in self._data_workspaces.get_loaded_workpace_names():
                 try:
                     if not ADS.doesExist(ws_name):
@@ -187,114 +72,7 @@ class FittingDataModel(object):
                         f"Failed to load file: {filename}. Error: {e}. \n Continuing loading of other files.")
             else:
                 logger.warning(f"File {ws_name} has already been loaded")
-        self.update_log_workspace_group()
-
-    def create_log_workspace_group(self):
-        # run information table
-        run_info = self.make_runinfo_table()
-        self._log_workspaces = GroupWorkspaces([run_info], OutputWorkspace='logs')
-        # a table per logs
-        logs = get_setting(output_settings.INTERFACES_SETTINGS_GROUP, output_settings.ENGINEERING_PREFIX, "logs")
-        if logs:
-            self._log_names = logs.split(',')
-            for log in self._log_names:
-                self.make_log_table(log)
-                self._log_workspaces.add(log)
-
-    def make_log_table(self, log):
-        ws_log = CreateEmptyTableWorkspace(OutputWorkspace=log)
-        ws_log.addColumn(type="float", name="avg")
-        ws_log.addColumn(type="float", name="stdev")
-        return ws_log
-
-    def make_runinfo_table(self):
-        run_info = CreateEmptyTableWorkspace()
-        run_info.addColumn(type="str", name="Instrument")
-        run_info.addColumn(type="int", name="Run")
-        run_info.addColumn(type="str", name="Bank")
-        run_info.addColumn(type="float", name="uAmps")
-        run_info.addColumn(type="str", name="Title")
-        return run_info
-
-    def update_log_workspace_group(self):
-        # both ws and name needed in event a ws is renamed and ws.name() is no longer correct
-
-        if not self._data_workspaces:
-            self.delete_logs()
-            return
-
-        if not self._log_workspaces:
-            self.create_log_workspace_group()
-        else:
-            for log in self._log_names:
-                if not ADS.doesExist(log):
-                    self.make_log_table(log)
-                    self._log_workspaces.add(log)
-            if not ADS.doesExist("run_info"):
-                self.make_runinfo_table()
-                self._log_workspaces.add("run_info")
-        # update log tables
-        self.remove_all_log_rows()
-        for irow, (ws_name, ws) in enumerate(self._data_workspaces.get_loaded_ws_dict().items()):
-            try:
-                self.add_log_to_table(ws_name, ws, irow)
-            except Exception as e:
-                logger.warning(f"Unable to output log workspaces for workspace {ws_name}: " + str(e))
-
-    def add_log_to_table(self, ws_name, ws, irow):
-        # both ws and name needed in event a ws is renamed and ws.name() is no longer correct
-        # make dict for run if doesn't exist
-        if ws_name not in self._log_values:
-            self._log_values[ws_name] = dict()
-        # add run info
-        run = ws.getRun()
-        row = [ws.getInstrument().getFullName(), ws.getRunNumber(), str(run.getProperty('bankid').value),
-               run.getProtonCharge(), ws.getTitle()]
-        self.write_table_row(ADS.retrieve("run_info"), row, irow)
-        # add log data - loop over existing log workspaces not logs in settings as these might have changed
-        currentRunLogs = [l.name for l in run.getLogData()]
-        nullLogValue = full(2, nan)  # default nan if can't read/average log data
-        if run.getProtonCharge() > 0 and "proton_charge" in currentRunLogs:
-            for log in self._log_names:
-                if log in self._log_values[ws_name]:
-                    avg, stdev = self._log_values[ws_name][log]  # already averaged
-                elif log in currentRunLogs:
-                    avg, stdev = AverageLogData(ws_name, LogName=log, FixZero=False)
-                else:
-                    avg, stdev = nullLogValue
-                self._log_values[ws_name][log] = [avg, stdev]  # update model dict (even if nan)
-        else:
-            self._log_values[ws_name] = {log: nullLogValue for log in self._log_names}
-            logger.warning(f"{ws.name()} does not contain a proton charge log - log values cannot be averaged.")
-
-        # write log values to table (nan if log could not be averaged)
-        for log, avg_and_stdev in self._log_values[ws_name].items():
-            self.write_table_row(ADS.retrieve(log), avg_and_stdev, irow)
-        self.update_log_group_name()
-
-    def remove_log_rows(self, row_numbers):
-        DeleteTableRows(TableWorkspace=self._log_workspaces, Rows=list(row_numbers))
-        self.update_log_group_name()
-
-    def remove_all_log_rows(self):
-        for ws in self._log_workspaces:
-            ws.setRowCount(0)
-
-    def delete_logs(self):
-        if self._log_workspaces:
-            ws_name = self._log_workspaces.name()
-            self._log_workspaces = None
-            DeleteWorkspace(ws_name)
-
-    def update_log_group_name(self):
-        run_info = ADS.retrieve('run_info')
-        if run_info.rowCount() > 0:
-            runs = run_info.column('Run')
-            name = f"{run_info.row(0)['Instrument']}_{min(runs)}-{max(runs)}_logs"
-            if not name == self._log_workspaces.name():
-                RenameWorkspace(InputWorkspace=self._log_workspaces.name(), OutputWorkspace=name)
-        else:
-            self.delete_logs()
+        self._sample_logs_workspace_group.update_log_workspace_group(self._data_workspaces)
 
     def get_loaded_ws_list(self):
         return list(self._data_workspaces.get_loaded_ws_dict().keys())
@@ -400,24 +178,25 @@ class FittingDataModel(object):
             if wsname in self._fit_results:
                 row = [wsname, self._fit_results[wsname]['costFunction'],
                        self._fit_results[wsname]['status'], self._fit_results[wsname]['model']]
-                self.write_table_row(model, row, iws)
+                write_table_row(model, row, iws)
             else:
-                self.write_table_row(model, ['', nan, ''], iws)
+                write_table_row(model, ['', nan, ''], iws)
         wslist += [model]
-        group_name = self._log_workspaces.name().split('_log')[0] + '_fits'
+        current_log_workspaces = self._sample_logs_workspace_group.get_log_workspaces()
+        group_name = current_log_workspaces.name().split('_log')[0] + '_fits'
         self._fit_workspaces = GroupWorkspaces(wslist, OutputWorkspace=group_name)
 
     # handle ADS remove. name workspace has already been deleted
     def remove_workspace(self, name):
         ws_loaded = self._data_workspaces.get(name, None)
         if ws_loaded:
-            bgsub_ws_name=self._data_workspaces[name].bgsub_ws_name
+            bgsub_ws_name = self._data_workspaces[name].bgsub_ws_name
             removed = self._data_workspaces.pop(name).loaded_ws
             # deleting bg sub workspace will generate further remove_workspace event so ensure this is done after
             # removing record from _data_workspaces to avoid circular call
             if bgsub_ws_name:
                 DeleteWorkspace(bgsub_ws_name)
-            self.update_log_workspace_group()
+            self._sample_logs_workspace_group.update_log_workspace_group(self._data_workspaces)
             return removed
         else:
             ws_loaded_name = self._data_workspaces.get_loaded_workspace_name_from_bgsub(name)
@@ -434,11 +213,12 @@ class FittingDataModel(object):
     def update_workspace_name(self, old_name, new_name):
         if new_name not in self.get_all_workspace_names():
             self._data_workspaces.rename(old_name, new_name)
-            if old_name in self._log_values:
-                self._log_values[new_name] = self._log_values.pop(old_name)
+            current_log_values = self._sample_logs_workspace_group.get_log_values()
+            if old_name in current_log_values:
+                self._sample_logs_workspace_group.update_log_value(new_key=new_name, old_key=old_name)
         else:
             logger.warning(f"There already exists a workspace with name {new_name}.")
-        self.update_log_workspace_group()
+            self._sample_logs_workspace_group.update_log_workspace_group(self._data_workspaces)
 
     # handle ADS clear
     def clear_workspaces(self):
@@ -446,9 +226,10 @@ class FittingDataModel(object):
         self.set_log_workspaces_none()
 
     def delete_workspaces(self):
-        if self._log_workspaces:
-            ws_name = self._log_workspaces.name()
-            self._log_workspaces = None
+        current_log_workspaces = self._sample_logs_workspace_group.get_log_workspaces()
+        if current_log_workspaces:
+            ws_name = current_log_workspaces.name()
+            self._sample_logs_workspace_group.clear_log_workspaces()
             DeleteWorkspace(ws_name)
         removed_ws_list = []
         for ws_name in self._data_workspaces.get_loaded_workpace_names():
@@ -462,7 +243,7 @@ class FittingDataModel(object):
         if removed.bgsub_ws:
             DeleteWorkspace(removed.bgsub_ws)
             removed_ws_list.append(removed.bgsub_ws)
-        self.update_log_workspace_group()
+            self._sample_logs_workspace_group.update_log_workspace_group(self._data_workspaces)
         return removed_ws_list
 
     def get_loaded_workspaces(self):
@@ -472,7 +253,8 @@ class FittingDataModel(object):
         return self._data_workspaces.get_loaded_workpace_names() + self._data_workspaces.get_bgsub_workpace_names()
 
     def get_log_workspaces_name(self):
-        return [ws.name() for ws in self._log_workspaces] if self._log_workspaces else ''
+        current_log_workspaces = self._sample_logs_workspace_group.get_log_workspaces()
+        return [ws.name() for ws in current_log_workspaces] if current_log_workspaces else ''
 
     def get_bgsub_workspaces(self):
         return self._data_workspaces.get_bgsub_ws_dict()
@@ -560,7 +342,7 @@ class FittingDataModel(object):
 
     def set_log_workspaces_none(self):
         # to be used in the event of Ads clear, as trying to reference the deleted grp ws results in an error
-        self._log_workspaces = None
+        self._sample_logs_workspace_group.clear_log_workspaces()
 
     def _convert_TOF_to_d(self, tof, ws_name):
         diff_consts = self._get_diff_constants(ws_name)
@@ -581,14 +363,3 @@ class FittingDataModel(object):
         si = ws.spectrumInfo()
         diff_consts = si.diffractometerConstants(0)  # output is a UnitParametersMap
         return diff_consts
-
-    @staticmethod
-    def write_table_row(ws_table, row, irow):
-        if irow > ws_table.rowCount() - 1:
-            ws_table.setRowCount(irow + 1)
-        [ws_table.setCell(irow, icol, row[icol]) for icol in range(0, len(row))]
-
-    @staticmethod
-    def _generate_workspace_name(filepath):
-        wsname = path.splitext(path.split(filepath)[1])[0]
-        return wsname
